+++ .gitignore
@@ -4,3 +4,4 @@ node_modules
 .test
 tmp
 coverage
+.nyc_output
