+++ make.js
@@ -1,12 +1,37 @@
 const b = require('substance-bundler')
 const path = require('path')
 const fs = require('fs')
-const fork = require('substance-bundler/extensions/fork')
 
 const DIST = 'dist/'
-const TMP ='tmp/'
+const TMP = 'tmp/'
 
-function _bundleANTLR4() {
+b.task('clean', () => {
+  b.rm(TMP)
+  b.rm(DIST)
+  b.rm('coverage')
+})
+  .describe('Cleans up temporary and build folders.')
+
+// ATM you we need to checkout the whole project and build a vendor bundle
+b.task('antlr4', _bundleANTLR4)
+  .describe('Prebundle antlr4.js')
+
+b.task('parser', _generateParser)
+  .describe('Generates the parser.\nRequires a clone of https://github/antlr4/antlr4.git as sibling folder')
+
+b.task('lib', ['parser'], _buildLib)
+  .describe('Builds the library into dist folder.')
+
+b.task('test:browser', ['lib'], () => {
+  _buildTestsBrowser()
+})
+  .describe('Builds the test-suite to be run from test/index.html.')
+
+b.task('default', ['clean', 'lib'])
+
+// HELPERS
+
+function _bundleANTLR4 () {
   b.browserify('../antlr4/runtime/JavaScript/src/antlr4/index', {
     dest: './vendor/antlr4.js',
     exports: ['default'],
@@ -14,7 +39,7 @@ function _bundleANTLR4() {
   })
 }
 
-function _generateParser() {
+function _generateParser () {
   // we can not build the parser without ANTLR4
   // still we don't fail so that travis is working (generated parser is checked in)
   if (!fs.existsSync('./.bin/antlr-4.6-complete.jar')) {
@@ -27,16 +52,17 @@ function _generateParser() {
   }
   b.custom('Generating parser', {
     src: './parser/Mini.g4',
-    dest: './parser/MiniParser.js',
+    dest: './tmp/MiniParser.js',
     execute: () => {
       const isWin = /^win/.test(process.platform)
       let cmd = 'java -jar ./.bin/antlr-4.6-complete.jar -Dlanguage=JavaScript -no-visitor'
       // WORKAROUND: antrl4 behaves differently under windows, i.e. it does not generate into folder 'parser'
       // thus we need to tell explicitely to do so
-      if (isWin) cmd += " -o parser"
+      if (isWin) cmd += ' -o tmp/parser'
+      else cmd += ' -o tmp'
       cmd += ' parser/Mini.g4'
       let exec = require('child_process').exec
-      return new Promise(function(resolve, reject) {
+      return new Promise(function (resolve, reject) {
         exec(cmd, (err) => {
           if (err) {
             reject(new Error(err))
@@ -44,40 +70,57 @@ function _generateParser() {
             resolve()
           }
         })
-      });
+      })
     }
   })
-  b.js('./parser/.parser.js', {
-    target: {
-      dest: './parser/parser.js',
-      format: 'es'
+  // NOTE: current versions commonjs and alias plugins
+  // are incompatible
+  // so we first bundle everything to es6 without the alias
+  // amd in a second step replacing the alias
+  b.js('./parser/parser.in.js', {
+    output: [{
+      file: './tmp/parser.js',
+      format: 'es',
+      sourcemap: false
+    }],
+    external: ['antlr4/index'],
+    commonjs: {
+      include: ['tmp/parser/**']
     },
+    cleanup: true
+  })
+  b.js('./tmp/parser.js', {
+    output: [{
+      file: './parser/parser.js',
+      format: 'es',
+      sourcemap: false
+    }],
     alias: {
       'antlr4/index': path.join(__dirname, '/vendor/antlr4.js')
-    },
-    commonjs: true
+    }
   })
 }
 
 // NOTE: this is needed when working on versions from github
-function _buildDeps() {
+function _buildDeps () {
   const subsDist = path.join(__dirname, 'node_modules/substance/dist')
-  if (!fs.existsSync(path.join(subsDist,'substance.js')) ||
+  if (!fs.existsSync(path.join(subsDist, 'substance.js')) ||
       !fs.existsSync(path.join(subsDist, 'substance.cjs.js'))) {
     b.make('substance')
   }
 }
 
-function _buildLib() {
+function _buildLib () {
   _buildDeps()
 
   b.js('index.js', {
-    targets: [{
-      dest: DIST+'stencila-mini.cjs.js',
+    output: [{
+      file: DIST + 'stencila-mini.cjs.js',
       format: 'cjs'
     }, {
-      dest: DIST+'stencila-mini.js',
-      format: 'umd', moduleName: 'stencilaMini',
+      file: DIST + 'stencila-mini.js',
+      format: 'umd',
+      name: 'stencilaMini',
       globals: {
         'substance': 'substance'
       }
@@ -86,89 +129,18 @@ function _buildLib() {
   })
 }
 
-function _buildTestsBrowser() {
+function _buildTestsBrowser () {
   b.js('./test/index.js', {
-    target: {
-      dest: TMP+'tests.js',
-      format: 'umd', moduleName: 'tests',
+    output: [{
+      file: TMP + 'tests.js',
+      format: 'umd',
+      name: 'tests',
       globals: {
         'substance': 'window.substance',
         'stencila-mini': 'window.stencilaMini',
         'substance-test': 'window.substanceTest'
       }
-    },
+    }],
     external: ['substance', 'stencila-mini', 'substance-test']
   })
 }
-
-function _buildTestsNodejs() {
-  b.js('./test/index.js', {
-    target: {
-      dest: TMP+'tests.cjs.js',
-      format: 'cjs'
-    },
-    alias: {
-      'stencila-mini': path.join(__dirname, 'index.js')
-    },
-    external: ['substance-test', 'substance']
-  })
-}
-
-function _buildTestsCov() {
-  b.js('./test/index.js', {
-    target: {
-      dest: TMP+'tests.cov.js',
-      format: 'cjs'
-    },
-    alias: {
-      'stencila-mini': path.join(__dirname, 'index.js')
-    },
-    external: ['substance-test', 'substance'],
-    istanbul: {
-      include: ['src/*.js']
-    }
-  })
-}
-
-b.task('clean', () => {
-  b.rm(TMP)
-  b.rm(DIST)
-  b.rm('coverage')
-})
-.describe('Cleans up temporary and build folders.')
-
-// ATM you we need to checkout the whole project and build a vendor bundle
-b.task('antlr4', _bundleANTLR4)
-.describe('Prebundle antlr4.js')
-
-b.task('parser', _generateParser)
-.describe('Generates the parser.\nRequires a clone of https://github/antlr4/antlr4.git as sibling folder')
-
-b.task('lib', ['parser'], _buildLib)
-.describe('Builds the library into dist folder.')
-
-b.task('test:browser', ['lib'], () => {
-  _buildTestsBrowser()
-})
-.describe('Builds the test-suite to be run from test/index.html.')
-
-b.task('test', ['lib'], () => {
-  const test = require.resolve('substance-test/bin/test')
-  const tests = './tmp/tests.cjs.js'
-  _buildTestsNodejs()
-  fork(b, test, tests, { verbose: true })
-})
-.describe('Runs the test-suite in node-js.')
-
-b.task('cover', ['lib'], () => {
-  const coverage = require.resolve('substance-test/bin/coverage')
-  const tests = './tmp/tests.cov.js'
-  _buildTestsCov()
-  fork(b, coverage, tests)
-})
-.describe('Runs the test-suite in node-js.')
-
-b.task('default', ['clean', 'lib'])
-
-b.setServerPort(5551)
-b.serve({ static: true, route: '/', folder: '.' })
