--- parser/Mini.tokens
@@ -1,60 +1,0 @@
-T__0=1
-T__1=2
-T__2=3
-T__3=4
-T__4=5
-T__5=6
-T__6=7
-T__7=8
-T__8=9
-T__9=10
-T__10=11
-T__11=12
-T__12=13
-T__13=14
-T__14=15
-T__15=16
-T__16=17
-T__17=18
-T__18=19
-T__19=20
-T__20=21
-T__21=22
-T__22=23
-T__23=24
-T__24=25
-T__25=26
-BOOLEAN=27
-ID=28
-INT=29
-FLOAT=30
-STRING=31
-EQ=32
-WS=33
-'.'=1
-'['=2
-']'=3
-'!'=4
-'+'=5
-'-'=6
-'^'=7
-'*'=8
-'/'=9
-'%'=10
-'<'=11
-'<='=12
-'>'=13
-'>='=14
-'=='=15
-'!='=16
-'&&'=17
-'||'=18
-'|'=19
-'('=20
-')'=21
-'()'=22
-','=23
-'{'=24
-':'=25
-'}'=26
-'='=32
