+++ parser/parser.js
@@ -3186,7 +3186,6 @@ function getCachedPredictionContext(context, contextCache, visited) {
 	return updated;
 }
 
-// ter's recursive version of Sam's getAllNodes()
 exports.merge = merge;
 exports.PredictionContext = PredictionContext;
 exports.PredictionContextCache = PredictionContextCache;
@@ -3668,7 +3667,7 @@ function arrayToString(a) {
 String.prototype.seed = String.prototype.seed || Math.round(Math.random() * Math.pow(2, 32));
 
 String.prototype.hashCode = function () {
-    var remainder, bytes, h1, h1b, c1, c1b, c2, c2b, k1, i,
+    var remainder, bytes, h1, h1b, c1, c2, k1, i,
         key = this.toString();
 
     remainder = key.length & 3; // key.length % 4
@@ -4073,7 +4072,6 @@ function titleCase(str) {
         return txt.charAt(0).toUpperCase() + txt.substr(1);
     });
 }
-
 function equalArrays(a, b)
 {
     if (!Array.isArray(a) || !Array.isArray(b))
@@ -4090,7 +4088,6 @@ function equalArrays(a, b)
     }
     return true;
 }
-
 exports.Hash = Hash;
 exports.Set = Set;
 exports.Map = Map;
@@ -4866,8 +4863,7 @@ ATNDeserializer.prototype.checkVersion = function() {
 ATNDeserializer.prototype.checkUUID = function() {
     var uuid = this.readUUID();
     if (SUPPORTED_UUIDS.indexOf(uuid)<0) {
-        throw ("Could not deserialize ATN with UUID: " + uuid +
-                        " (expected " + SERIALIZED_UUID + " or a legacy UUID).", uuid, SERIALIZED_UUID);
+        throw (SERIALIZED_UUID);
     }
     this.uuid = uuid;
 };
@@ -5439,74 +5435,6 @@ ATNSimulator.prototype.getCachedContext = function(context) {
 exports.ATNSimulator = ATNSimulator;
 
 },{"./../PredictionContext":11,"./../dfa/DFAState":34,"./ATNConfigSet":18}],22:[function(require,module,exports){
-//
-/* Copyright (c) 2012-2016 The ANTLR Project. All rights reserved.
- * Use of this file is governed by the BSD 3-clause license that
- * can be found in the LICENSE.txt file in the project root.
- */
-//
-
-// The following images show the relation of states and
-// {@link ATNState//transitions} for various grammar constructs.
-//
-// <ul>
-//
-// <li>Solid edges marked with an &//0949; indicate a required
-// {@link EpsilonTransition}.</li>
-//
-// <li>Dashed edges indicate locations where any transition derived from
-// {@link Transition} might appear.</li>
-//
-// <li>Dashed nodes are place holders for either a sequence of linked
-// {@link BasicState} states or the inclusion of a block representing a nested
-// construct in one of the forms below.</li>
-//
-// <li>Nodes showing multiple outgoing alternatives with a {@code ...} support
-// any number of alternatives (one or more). Nodes without the {@code ...} only
-// support the exact number of alternatives shown in the diagram.</li>
-//
-// </ul>
-//
-// <h2>Basic Blocks</h2>
-//
-// <h3>Rule</h3>
-//
-// <embed src="images/Rule.svg" type="image/svg+xml"/>
-//
-// <h3>Block of 1 or more alternatives</h3>
-//
-// <embed src="images/Block.svg" type="image/svg+xml"/>
-//
-// <h2>Greedy Loops</h2>
-//
-// <h3>Greedy Closure: {@code (...)*}</h3>
-//
-// <embed src="images/ClosureGreedy.svg" type="image/svg+xml"/>
-//
-// <h3>Greedy Positive Closure: {@code (...)+}</h3>
-//
-// <embed src="images/PositiveClosureGreedy.svg" type="image/svg+xml"/>
-//
-// <h3>Greedy Optional: {@code (...)?}</h3>
-//
-// <embed src="images/OptionalGreedy.svg" type="image/svg+xml"/>
-//
-// <h2>Non-Greedy Loops</h2>
-//
-// <h3>Non-Greedy Closure: {@code (...)*?}</h3>
-//
-// <embed src="images/ClosureNonGreedy.svg" type="image/svg+xml"/>
-//
-// <h3>Non-Greedy Positive Closure: {@code (...)+?}</h3>
-//
-// <embed src="images/PositiveClosureNonGreedy.svg" type="image/svg+xml"/>
-//
-// <h3>Non-Greedy Optional: {@code (...)??}</h3>
-//
-// <embed src="images/OptionalNonGreedy.svg" type="image/svg+xml"/>
-//
-
-var INITIAL_NUM_TRANSITIONS = 4;
 
 function ATNState() {
     // Which ATN are we in?
@@ -11360,8 +11288,6 @@ function RecognitionException(params) {
 	Error.call(this);
 	if (!!Error.captureStackTrace) {
         Error.captureStackTrace(this, RecognitionException);
-	} else {
-		var stack = new Error().stack;
 	}
 	this.message = params.message;
     this.recognizer = params.recognizer;
@@ -11556,7 +11482,6 @@ exports.Utils = require('./Utils');
 /*! https://mths.be/codepointat v0.2.0 by @mathias */
 if (!String.prototype.codePointAt) {
 	(function() {
-		'use strict'; // needed to support `apply`/`call` with `undefined`/`null`
 		var defineProperty = (function() {
 			// IE 8 only supports `Object.defineProperty` on DOM elements
 			try {
@@ -12058,18 +11983,6 @@ exports.ParseTreeWalker = Tree.ParseTreeWalker;
 
 },{}]},{},[41])(41);
 
-
-var antlr4 = Object.freeze({
-	default: _exports
-});
-
-var antlr4$1 = ( antlr4 && _exports ) || antlr4;
-
-// Generated from parser/Mini.g4 by ANTLR 4.6
-// jshint ignore: start
-
-
-
 var serializedATN = ["\u0003\u0430\ud6d1\u8206\uad2d\u4417\uaef1\u8d80\uaadd",
     "\u0002#\u00c1\b\u0001\u0004\u0002\t\u0002\u0004\u0003\t\u0003\u0004",
     "\u0004\t\u0004\u0004\u0005\t\u0005\u0004\u0006\t\u0006\u0004\u0007\t",
@@ -12196,58 +12109,58 @@ var serializedATN = ["\u0003\u0430\ud6d1\u8206\uad2d\u4417\uaef1\u8d80\uaadd",
     "\u00b0\u00b2\u00b6\u00bd\u0003\b\u0002\u0002"].join("");
 
 
-var atn = new antlr4$1.atn.ATNDeserializer().deserialize(serializedATN);
+var atn = new _exports.atn.ATNDeserializer().deserialize(serializedATN);
 
-var decisionsToDFA = atn.decisionToState.map( function(ds, index) { return new antlr4$1.dfa.DFA(ds, index); });
+var decisionsToDFA = atn.decisionToState.map( function(ds, index) { return new _exports.dfa.DFA(ds, index); });
 
-function MiniLexer$1(input) {
-	antlr4$1.Lexer.call(this, input);
-    this._interp = new antlr4$1.atn.LexerATNSimulator(this, atn, decisionsToDFA, new antlr4$1.PredictionContextCache());
+function MiniLexer(input) {
+	_exports.Lexer.call(this, input);
+    this._interp = new _exports.atn.LexerATNSimulator(this, atn, decisionsToDFA, new _exports.PredictionContextCache());
     return this;
 }
 
-MiniLexer$1.prototype = Object.create(antlr4$1.Lexer.prototype);
-MiniLexer$1.prototype.constructor = MiniLexer$1;
+MiniLexer.prototype = Object.create(_exports.Lexer.prototype);
+MiniLexer.prototype.constructor = MiniLexer;
 
-MiniLexer$1.EOF = antlr4$1.Token.EOF;
-MiniLexer$1.T__0 = 1;
-MiniLexer$1.T__1 = 2;
-MiniLexer$1.T__2 = 3;
-MiniLexer$1.T__3 = 4;
-MiniLexer$1.T__4 = 5;
-MiniLexer$1.T__5 = 6;
-MiniLexer$1.T__6 = 7;
-MiniLexer$1.T__7 = 8;
-MiniLexer$1.T__8 = 9;
-MiniLexer$1.T__9 = 10;
-MiniLexer$1.T__10 = 11;
-MiniLexer$1.T__11 = 12;
-MiniLexer$1.T__12 = 13;
-MiniLexer$1.T__13 = 14;
-MiniLexer$1.T__14 = 15;
-MiniLexer$1.T__15 = 16;
-MiniLexer$1.T__16 = 17;
-MiniLexer$1.T__17 = 18;
-MiniLexer$1.T__18 = 19;
-MiniLexer$1.T__19 = 20;
-MiniLexer$1.T__20 = 21;
-MiniLexer$1.T__21 = 22;
-MiniLexer$1.T__22 = 23;
-MiniLexer$1.T__23 = 24;
-MiniLexer$1.T__24 = 25;
-MiniLexer$1.T__25 = 26;
-MiniLexer$1.BOOLEAN = 27;
-MiniLexer$1.ID = 28;
-MiniLexer$1.INT = 29;
-MiniLexer$1.FLOAT = 30;
-MiniLexer$1.STRING = 31;
-MiniLexer$1.EQ = 32;
-MiniLexer$1.WS = 33;
+MiniLexer.EOF = _exports.Token.EOF;
+MiniLexer.T__0 = 1;
+MiniLexer.T__1 = 2;
+MiniLexer.T__2 = 3;
+MiniLexer.T__3 = 4;
+MiniLexer.T__4 = 5;
+MiniLexer.T__5 = 6;
+MiniLexer.T__6 = 7;
+MiniLexer.T__7 = 8;
+MiniLexer.T__8 = 9;
+MiniLexer.T__9 = 10;
+MiniLexer.T__10 = 11;
+MiniLexer.T__11 = 12;
+MiniLexer.T__12 = 13;
+MiniLexer.T__13 = 14;
+MiniLexer.T__14 = 15;
+MiniLexer.T__15 = 16;
+MiniLexer.T__16 = 17;
+MiniLexer.T__17 = 18;
+MiniLexer.T__18 = 19;
+MiniLexer.T__19 = 20;
+MiniLexer.T__20 = 21;
+MiniLexer.T__21 = 22;
+MiniLexer.T__22 = 23;
+MiniLexer.T__23 = 24;
+MiniLexer.T__24 = 25;
+MiniLexer.T__25 = 26;
+MiniLexer.BOOLEAN = 27;
+MiniLexer.ID = 28;
+MiniLexer.INT = 29;
+MiniLexer.FLOAT = 30;
+MiniLexer.STRING = 31;
+MiniLexer.EQ = 32;
+MiniLexer.WS = 33;
 
 
-MiniLexer$1.prototype.modeNames = [ "DEFAULT_MODE" ];
+MiniLexer.prototype.modeNames = [ "DEFAULT_MODE" ];
 
-MiniLexer$1.prototype.literalNames = [ null, "'.'", "'['", "']'", "'!'", "'+'", 
+MiniLexer.prototype.literalNames = [ null, "'.'", "'['", "']'", "'!'", "'+'", 
                                      "'-'", "'^'", "'*'", "'/'", "'%'", 
                                      "'<'", "'<='", "'>'", "'>='", "'=='", 
                                      "'!='", "'&&'", "'||'", "'|'", "'('", 
@@ -12255,14 +12168,14 @@ MiniLexer$1.prototype.literalNames = [ null, "'.'", "'['", "']'", "'!'", "'+'",
                                      "'}'", null, null, null, null, null, 
                                      "'='" ];
 
-MiniLexer$1.prototype.symbolicNames = [ null, null, null, null, null, null, 
+MiniLexer.prototype.symbolicNames = [ null, null, null, null, null, null, 
                                       null, null, null, null, null, null, 
                                       null, null, null, null, null, null, 
                                       null, null, null, null, null, null, 
                                       null, null, null, "BOOLEAN", "ID", 
                                       "INT", "FLOAT", "STRING", "EQ", "WS" ];
 
-MiniLexer$1.prototype.ruleNames = [ "T__0", "T__1", "T__2", "T__3", "T__4", 
+MiniLexer.prototype.ruleNames = [ "T__0", "T__1", "T__2", "T__3", "T__4", 
                                   "T__5", "T__6", "T__7", "T__8", "T__9", 
                                   "T__10", "T__11", "T__12", "T__13", "T__14", 
                                   "T__15", "T__16", "T__17", "T__18", "T__19", 
@@ -12270,157 +12183,150 @@ MiniLexer$1.prototype.ruleNames = [ "T__0", "T__1", "T__2", "T__3", "T__4",
                                   "T__25", "BOOLEAN", "ID", "INT", "FLOAT", 
                                   "STRING", "EQ", "WS" ];
 
-MiniLexer$1.prototype.grammarFileName = "Mini.g4";
+MiniLexer.prototype.grammarFileName = "Mini.g4";
 
 
 
-var MiniLexer_2 = MiniLexer$1;
+var MiniLexer_2 = MiniLexer;
 
 var MiniLexer_1 = {
 	MiniLexer: MiniLexer_2
 };
 
-// Generated from parser/Mini.g4 by ANTLR 4.6
-// jshint ignore: start
-
-
-// This class defines a complete listener for a parse tree produced by MiniParser.
-function MiniListener$2() {
-	antlr4$1.tree.ParseTreeListener.call(this);
+function MiniListener() {
+	_exports.tree.ParseTreeListener.call(this);
 	return this;
 }
 
-MiniListener$2.prototype = Object.create(antlr4$1.tree.ParseTreeListener.prototype);
-MiniListener$2.prototype.constructor = MiniListener$2;
+MiniListener.prototype = Object.create(_exports.tree.ParseTreeListener.prototype);
+MiniListener.prototype.constructor = MiniListener;
 
-// Enter a parse tree produced by MiniParser#mini.
-MiniListener$2.prototype.enterMini = function(ctx) {
-};
 
-// Exit a parse tree produced by MiniParser#mini.
-MiniListener$2.prototype.exitMini = function(ctx) {
+MiniListener.prototype.enterMini = function(ctx) {
 };
 
 
-// Enter a parse tree produced by MiniParser#mainExpr.
-MiniListener$2.prototype.enterMainExpr = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#mainExpr.
-MiniListener$2.prototype.exitMainExpr = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#expr.
-MiniListener$2.prototype.enterExpr = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#expr.
-MiniListener$2.prototype.exitExpr = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#function_call.
-MiniListener$2.prototype.enterFunction_call = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#function_call.
-MiniListener$2.prototype.exitFunction_call = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#number.
-MiniListener$2.prototype.enterNumber = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#number.
-MiniListener$2.prototype.exitNumber = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#seq.
-MiniListener$2.prototype.enterSeq = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#seq.
-MiniListener$2.prototype.exitSeq = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#id_seq.
-MiniListener$2.prototype.enterId_seq = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#id_seq.
-MiniListener$2.prototype.exitId_seq = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#call_arguments.
-MiniListener$2.prototype.enterCall_arguments = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#call_arguments.
-MiniListener$2.prototype.exitCall_arguments = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#positional_arguments.
-MiniListener$2.prototype.enterPositional_arguments = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#positional_arguments.
-MiniListener$2.prototype.exitPositional_arguments = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#named_arguments.
-MiniListener$2.prototype.enterNamed_arguments = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#named_arguments.
-MiniListener$2.prototype.exitNamed_arguments = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#named_argument.
-MiniListener$2.prototype.enterNamed_argument = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#named_argument.
-MiniListener$2.prototype.exitNamed_argument = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#array.
-MiniListener$2.prototype.enterArray = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#array.
-MiniListener$2.prototype.exitArray = function(ctx) {
-};
-
-
-// Enter a parse tree produced by MiniParser#object.
-MiniListener$2.prototype.enterObject = function(ctx) {
-};
-
-// Exit a parse tree produced by MiniParser#object.
-MiniListener$2.prototype.exitObject = function(ctx) {
+MiniListener.prototype.exitMini = function(ctx) {
 };
 
 
 
-var MiniListener_2 = MiniListener$2;
+MiniListener.prototype.enterMainExpr = function(ctx) {
+};
+
+
+MiniListener.prototype.exitMainExpr = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterExpr = function(ctx) {
+};
+
+
+MiniListener.prototype.exitExpr = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterFunction_call = function(ctx) {
+};
+
+
+MiniListener.prototype.exitFunction_call = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterNumber = function(ctx) {
+};
+
+
+MiniListener.prototype.exitNumber = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterSeq = function(ctx) {
+};
+
+
+MiniListener.prototype.exitSeq = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterId_seq = function(ctx) {
+};
+
+
+MiniListener.prototype.exitId_seq = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterCall_arguments = function(ctx) {
+};
+
+
+MiniListener.prototype.exitCall_arguments = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterPositional_arguments = function(ctx) {
+};
+
+
+MiniListener.prototype.exitPositional_arguments = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterNamed_arguments = function(ctx) {
+};
+
+
+MiniListener.prototype.exitNamed_arguments = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterNamed_argument = function(ctx) {
+};
+
+
+MiniListener.prototype.exitNamed_argument = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterArray = function(ctx) {
+};
+
+
+MiniListener.prototype.exitArray = function(ctx) {
+};
+
+
+
+MiniListener.prototype.enterObject = function(ctx) {
+};
+
+
+MiniListener.prototype.exitObject = function(ctx) {
+};
+
+
+
+var MiniListener_2 = MiniListener;
 
 var MiniListener_1 = {
 	MiniListener: MiniListener_2
 };
 
-// Generated from parser/Mini.g4 by ANTLR 4.6
-// jshint ignore: start
-
 var MiniListener$1 = MiniListener_1.MiniListener;
+
 var serializedATN$1 = ["\u0003\u0430\ud6d1\u8206\uad2d\u4417\uaef1\u8d80\uaadd",
     "\u0003#\u00de\u0004\u0002\t\u0002\u0004\u0003\t\u0003\u0004\u0004\t",
     "\u0004\u0004\u0005\t\u0005\u0004\u0006\t\u0006\u0004\u0007\t\u0007\u0004",
@@ -12565,11 +12471,11 @@ var serializedATN$1 = ["\u0003\u0430\ud6d1\u8206\uad2d\u4417\uaef1\u8d80\uaadd",
     "\u0090\u0097\u009f\u00a9\u00ae\u00b2\u00b8\u00bc\u00c6\u00d5\u00d8"].join("");
 
 
-var atn$1 = new antlr4$1.atn.ATNDeserializer().deserialize(serializedATN$1);
+var atn$1 = new _exports.atn.ATNDeserializer().deserialize(serializedATN$1);
 
-var decisionsToDFA$1 = atn$1.decisionToState.map( function(ds, index) { return new antlr4$1.dfa.DFA(ds, index); });
+var decisionsToDFA$1 = atn$1.decisionToState.map( function(ds, index) { return new _exports.dfa.DFA(ds, index); });
 
-var sharedContextCache = new antlr4$1.PredictionContextCache();
+var sharedContextCache = new _exports.PredictionContextCache();
 
 var literalNames = [ null, "'.'", "'['", "']'", "'!'", "'+'", "'-'", "'^'", 
                      "'*'", "'/'", "'%'", "'<'", "'<='", "'>'", "'>='", 
@@ -12586,72 +12492,72 @@ var ruleNames =  [ "mini", "mainExpr", "expr", "function_call", "number",
                    "seq", "id_seq", "call_arguments", "positional_arguments", 
                    "named_arguments", "named_argument", "array", "object" ];
 
-function MiniParser$1 (input) {
-	antlr4$1.Parser.call(this, input);
-    this._interp = new antlr4$1.atn.ParserATNSimulator(this, atn$1, decisionsToDFA$1, sharedContextCache);
+function MiniParser (input) {
+	_exports.Parser.call(this, input);
+    this._interp = new _exports.atn.ParserATNSimulator(this, atn$1, decisionsToDFA$1, sharedContextCache);
     this.ruleNames = ruleNames;
     this.literalNames = literalNames;
     this.symbolicNames = symbolicNames;
     return this;
 }
 
-MiniParser$1.prototype = Object.create(antlr4$1.Parser.prototype);
-MiniParser$1.prototype.constructor = MiniParser$1;
+MiniParser.prototype = Object.create(_exports.Parser.prototype);
+MiniParser.prototype.constructor = MiniParser;
 
-Object.defineProperty(MiniParser$1.prototype, "atn", {
+Object.defineProperty(MiniParser.prototype, "atn", {
 	get : function() {
 		return atn$1;
 	}
 });
 
-MiniParser$1.EOF = antlr4$1.Token.EOF;
-MiniParser$1.T__0 = 1;
-MiniParser$1.T__1 = 2;
-MiniParser$1.T__2 = 3;
-MiniParser$1.T__3 = 4;
-MiniParser$1.T__4 = 5;
-MiniParser$1.T__5 = 6;
-MiniParser$1.T__6 = 7;
-MiniParser$1.T__7 = 8;
-MiniParser$1.T__8 = 9;
-MiniParser$1.T__9 = 10;
-MiniParser$1.T__10 = 11;
-MiniParser$1.T__11 = 12;
-MiniParser$1.T__12 = 13;
-MiniParser$1.T__13 = 14;
-MiniParser$1.T__14 = 15;
-MiniParser$1.T__15 = 16;
-MiniParser$1.T__16 = 17;
-MiniParser$1.T__17 = 18;
-MiniParser$1.T__18 = 19;
-MiniParser$1.T__19 = 20;
-MiniParser$1.T__20 = 21;
-MiniParser$1.T__21 = 22;
-MiniParser$1.T__22 = 23;
-MiniParser$1.T__23 = 24;
-MiniParser$1.T__24 = 25;
-MiniParser$1.T__25 = 26;
-MiniParser$1.BOOLEAN = 27;
-MiniParser$1.ID = 28;
-MiniParser$1.INT = 29;
-MiniParser$1.FLOAT = 30;
-MiniParser$1.STRING = 31;
-MiniParser$1.EQ = 32;
-MiniParser$1.WS = 33;
+MiniParser.EOF = _exports.Token.EOF;
+MiniParser.T__0 = 1;
+MiniParser.T__1 = 2;
+MiniParser.T__2 = 3;
+MiniParser.T__3 = 4;
+MiniParser.T__4 = 5;
+MiniParser.T__5 = 6;
+MiniParser.T__6 = 7;
+MiniParser.T__7 = 8;
+MiniParser.T__8 = 9;
+MiniParser.T__9 = 10;
+MiniParser.T__10 = 11;
+MiniParser.T__11 = 12;
+MiniParser.T__12 = 13;
+MiniParser.T__13 = 14;
+MiniParser.T__14 = 15;
+MiniParser.T__15 = 16;
+MiniParser.T__16 = 17;
+MiniParser.T__17 = 18;
+MiniParser.T__18 = 19;
+MiniParser.T__19 = 20;
+MiniParser.T__20 = 21;
+MiniParser.T__21 = 22;
+MiniParser.T__22 = 23;
+MiniParser.T__23 = 24;
+MiniParser.T__24 = 25;
+MiniParser.T__25 = 26;
+MiniParser.BOOLEAN = 27;
+MiniParser.ID = 28;
+MiniParser.INT = 29;
+MiniParser.FLOAT = 30;
+MiniParser.STRING = 31;
+MiniParser.EQ = 32;
+MiniParser.WS = 33;
 
-MiniParser$1.RULE_mini = 0;
-MiniParser$1.RULE_mainExpr = 1;
-MiniParser$1.RULE_expr = 2;
-MiniParser$1.RULE_function_call = 3;
-MiniParser$1.RULE_number = 4;
-MiniParser$1.RULE_seq = 5;
-MiniParser$1.RULE_id_seq = 6;
-MiniParser$1.RULE_call_arguments = 7;
-MiniParser$1.RULE_positional_arguments = 8;
-MiniParser$1.RULE_named_arguments = 9;
-MiniParser$1.RULE_named_argument = 10;
-MiniParser$1.RULE_array = 11;
-MiniParser$1.RULE_object = 12;
+MiniParser.RULE_mini = 0;
+MiniParser.RULE_mainExpr = 1;
+MiniParser.RULE_expr = 2;
+MiniParser.RULE_function_call = 3;
+MiniParser.RULE_number = 4;
+MiniParser.RULE_seq = 5;
+MiniParser.RULE_id_seq = 6;
+MiniParser.RULE_call_arguments = 7;
+MiniParser.RULE_positional_arguments = 8;
+MiniParser.RULE_named_arguments = 9;
+MiniParser.RULE_named_argument = 10;
+MiniParser.RULE_array = 11;
+MiniParser.RULE_object = 12;
 
 function MiniContext(parser, parent, invokingState) {
 	if(parent===undefined) {
@@ -12660,13 +12566,13 @@ function MiniContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_mini;
+    this.ruleIndex = MiniParser.RULE_mini;
     return this;
 }
 
-MiniContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+MiniContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 MiniContext.prototype.constructor = MiniContext;
 
 MiniContext.prototype.mainExpr = function() {
@@ -12674,15 +12580,15 @@ MiniContext.prototype.mainExpr = function() {
 };
 
 MiniContext.prototype.EOF = function() {
-    return this.getToken(MiniParser$1.EOF, 0);
+    return this.getToken(MiniParser.EOF, 0);
 };
 
 MiniContext.prototype.ID = function() {
-    return this.getToken(MiniParser$1.ID, 0);
+    return this.getToken(MiniParser.ID, 0);
 };
 
 MiniContext.prototype.EQ = function() {
-    return this.getToken(MiniParser$1.EQ, 0);
+    return this.getToken(MiniParser.EQ, 0);
 };
 
 MiniContext.prototype.enterRule = function(listener) {
@@ -12700,12 +12606,12 @@ MiniContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.MiniContext = MiniContext;
+MiniParser.MiniContext = MiniContext;
 
-MiniParser$1.prototype.mini = function() {
+MiniParser.prototype.mini = function() {
 
     var localctx = new MiniContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 0, MiniParser$1.RULE_mini);
+    this.enterRule(localctx, 0, MiniParser.RULE_mini);
     try {
         this.state = 36;
         this._errHandler.sync(this);
@@ -12716,26 +12622,26 @@ MiniParser$1.prototype.mini = function() {
             this.state = 26;
             this.mainExpr();
             this.state = 27;
-            this.match(MiniParser$1.EOF);
+            this.match(MiniParser.EOF);
              localctx.type = 'evaluation'; 
             break;
 
         case 2:
             this.enterOuterAlt(localctx, 2);
             this.state = 30;
-            this.match(MiniParser$1.ID);
+            this.match(MiniParser.ID);
             this.state = 31;
-            this.match(MiniParser$1.EQ);
+            this.match(MiniParser.EQ);
             this.state = 32;
             this.mainExpr();
             this.state = 33;
-            this.match(MiniParser$1.EOF);
+            this.match(MiniParser.EOF);
              localctx.type = 'definition';
             break;
 
         }
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -12755,13 +12661,13 @@ function MainExprContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_mainExpr;
+    this.ruleIndex = MiniParser.RULE_mainExpr;
     return this;
 }
 
-MainExprContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+MainExprContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 MainExprContext.prototype.constructor = MainExprContext;
 
 MainExprContext.prototype.expr = function() {
@@ -12783,19 +12689,19 @@ MainExprContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.MainExprContext = MainExprContext;
+MiniParser.MainExprContext = MainExprContext;
 
-MiniParser$1.prototype.mainExpr = function() {
+MiniParser.prototype.mainExpr = function() {
 
     var localctx = new MainExprContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 2, MiniParser$1.RULE_mainExpr);
+    this.enterRule(localctx, 2, MiniParser.RULE_mainExpr);
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 38;
         this.expr(0);
          localctx.type = 'simple'; 
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -12815,14 +12721,14 @@ function ExprContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_expr;
-    this.op = null; // Token
+    this.ruleIndex = MiniParser.RULE_expr;
+    this.op = null; 
     return this;
 }
 
-ExprContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+ExprContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 ExprContext.prototype.constructor = ExprContext;
 
 ExprContext.prototype.expr = function(i) {
@@ -12837,7 +12743,7 @@ ExprContext.prototype.expr = function(i) {
 };
 
 ExprContext.prototype.BOOLEAN = function() {
-    return this.getToken(MiniParser$1.BOOLEAN, 0);
+    return this.getToken(MiniParser.BOOLEAN, 0);
 };
 
 ExprContext.prototype.number = function() {
@@ -12845,7 +12751,7 @@ ExprContext.prototype.number = function() {
 };
 
 ExprContext.prototype.ID = function() {
-    return this.getToken(MiniParser$1.ID, 0);
+    return this.getToken(MiniParser.ID, 0);
 };
 
 ExprContext.prototype.function_call = function() {
@@ -12861,7 +12767,7 @@ ExprContext.prototype.object = function() {
 };
 
 ExprContext.prototype.STRING = function() {
-    return this.getToken(MiniParser$1.STRING, 0);
+    return this.getToken(MiniParser.STRING, 0);
 };
 
 ExprContext.prototype.enterRule = function(listener) {
@@ -12878,17 +12784,16 @@ ExprContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.prototype.expr = function(_p) {
+MiniParser.prototype.expr = function(_p) {
 	if(_p===undefined) {
 	    _p = 0;
 	}
     var _parentctx = this._ctx;
     var _parentState = this.state;
     var localctx = new ExprContext(this, this._ctx, _parentState);
-    var _prevctx = localctx;
     var _startState = 4;
-    this.enterRecursionRule(localctx, 4, MiniParser$1.RULE_expr, _p);
-    var _la = 0; // Token type
+    this.enterRecursionRule(localctx, 4, MiniParser.RULE_expr, _p);
+    var _la = 0; 
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 70;
@@ -12899,7 +12804,7 @@ MiniParser$1.prototype.expr = function(_p) {
             this.state = 42;
             localctx.op = this._input.LT(1);
             _la = this._input.LA(1);
-            if(!((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser$1.T__3) | (1 << MiniParser$1.T__4) | (1 << MiniParser$1.T__5))) !== 0))) {
+            if(!((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser.T__3) | (1 << MiniParser.T__4) | (1 << MiniParser.T__5))) !== 0))) {
                 localctx.op = this._errHandler.recoverInline(this);
             }
             else {
@@ -12917,7 +12822,7 @@ MiniParser$1.prototype.expr = function(_p) {
 
         case 2:
             this.state = 46;
-            this.match(MiniParser$1.BOOLEAN);
+            this.match(MiniParser.BOOLEAN);
              localctx.type = 'boolean'; 
             break;
 
@@ -12929,10 +12834,10 @@ MiniParser$1.prototype.expr = function(_p) {
 
         case 4:
             this.state = 51;
-            this.match(MiniParser$1.ID);
+            this.match(MiniParser.ID);
             this.state = 52;
-            if (!( this._input.LA(1) !== MiniParser$1.EQ)) {
-                throw new antlr4$1.error.FailedPredicateException(this, "this._input.LA(1) !== MiniParser.EQ");
+            if (!( this._input.LA(1) !== MiniParser.EQ)) {
+                throw new _exports.error.FailedPredicateException(this, "this._input.LA(1) !== MiniParser.EQ");
             }
              localctx.type = 'var'; 
             break;
@@ -12945,11 +12850,11 @@ MiniParser$1.prototype.expr = function(_p) {
 
         case 6:
             this.state = 57;
-            this.match(MiniParser$1.T__19);
+            this.match(MiniParser.T__19);
             this.state = 58;
             this.expr(0);
             this.state = 59;
-            this.match(MiniParser$1.T__20);
+            this.match(MiniParser.T__20);
              localctx.type = 'group'; 
             break;
 
@@ -12967,7 +12872,7 @@ MiniParser$1.prototype.expr = function(_p) {
 
         case 9:
             this.state = 68;
-            this.match(MiniParser$1.STRING);
+            this.match(MiniParser.STRING);
              localctx.type = 'string'; 
             break;
 
@@ -12976,25 +12881,24 @@ MiniParser$1.prototype.expr = function(_p) {
         this.state = 124;
         this._errHandler.sync(this);
         var _alt = this._interp.adaptivePredict(this._input,3,this._ctx);
-        while(_alt!=2 && _alt!=antlr4$1.atn.ATN.INVALID_ALT_NUMBER) {
+        while(_alt!=2 && _alt!=_exports.atn.ATN.INVALID_ALT_NUMBER) {
             if(_alt===1) {
                 if(this._parseListeners!==null) {
                     this.triggerExitRuleEvent();
                 }
-                _prevctx = localctx;
                 this.state = 122;
                 this._errHandler.sync(this);
                 var la_ = this._interp.adaptivePredict(this._input,2,this._ctx);
                 switch(la_) {
                 case 1:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 72;
                     if (!( this.precpred(this._ctx, 16))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 16)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 16)");
                     }
                     this.state = 73;
-                    this.match(MiniParser$1.T__6);
+                    this.match(MiniParser.T__6);
                     this.state = 74;
                     this.expr(17);
 
@@ -13004,15 +12908,15 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 2:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 77;
                     if (!( this.precpred(this._ctx, 15))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 15)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 15)");
                     }
                     this.state = 78;
                     localctx.op = this._input.LT(1);
                     _la = this._input.LA(1);
-                    if(!((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser$1.T__7) | (1 << MiniParser$1.T__8) | (1 << MiniParser$1.T__9))) !== 0))) {
+                    if(!((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser.T__7) | (1 << MiniParser.T__8) | (1 << MiniParser.T__9))) !== 0))) {
                         localctx.op = this._errHandler.recoverInline(this);
                     }
                     else {
@@ -13030,15 +12934,15 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 3:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 82;
                     if (!( this.precpred(this._ctx, 14))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 14)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 14)");
                     }
                     this.state = 83;
                     localctx.op = this._input.LT(1);
                     _la = this._input.LA(1);
-                    if(!(_la===MiniParser$1.T__4 || _la===MiniParser$1.T__5)) {
+                    if(!(_la===MiniParser.T__4 || _la===MiniParser.T__5)) {
                         localctx.op = this._errHandler.recoverInline(this);
                     }
                     else {
@@ -13054,15 +12958,15 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 4:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 87;
                     if (!( this.precpred(this._ctx, 13))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 13)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 13)");
                     }
                     this.state = 88;
                     localctx.op = this._input.LT(1);
                     _la = this._input.LA(1);
-                    if(!((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser$1.T__10) | (1 << MiniParser$1.T__11) | (1 << MiniParser$1.T__12) | (1 << MiniParser$1.T__13))) !== 0))) {
+                    if(!((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser.T__10) | (1 << MiniParser.T__11) | (1 << MiniParser.T__12) | (1 << MiniParser.T__13))) !== 0))) {
                         localctx.op = this._errHandler.recoverInline(this);
                     }
                     else {
@@ -13081,15 +12985,15 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 5:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 92;
                     if (!( this.precpred(this._ctx, 12))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 12)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 12)");
                     }
                     this.state = 93;
                     localctx.op = this._input.LT(1);
                     _la = this._input.LA(1);
-                    if(!(_la===MiniParser$1.T__14 || _la===MiniParser$1.T__15)) {
+                    if(!(_la===MiniParser.T__14 || _la===MiniParser.T__15)) {
                         localctx.op = this._errHandler.recoverInline(this);
                     }
                     else {
@@ -13105,13 +13009,13 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 6:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 97;
                     if (!( this.precpred(this._ctx, 11))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 11)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 11)");
                     }
                     this.state = 98;
-                    this.match(MiniParser$1.T__16);
+                    this.match(MiniParser.T__16);
                     this.state = 99;
                     this.expr(12);
                      localctx.type = 'and'; 
@@ -13119,13 +13023,13 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 7:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 102;
                     if (!( this.precpred(this._ctx, 10))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 10)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 10)");
                     }
                     this.state = 103;
-                    this.match(MiniParser$1.T__17);
+                    this.match(MiniParser.T__17);
                     this.state = 104;
                     this.expr(11);
                      localctx.type = 'or'; 
@@ -13133,15 +13037,15 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 8:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 107;
                     if (!( this.precpred(this._ctx, 19))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 19)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 19)");
                     }
                     this.state = 108;
-                    this.match(MiniParser$1.T__0);
+                    this.match(MiniParser.T__0);
                     this.state = 109;
-                    this.match(MiniParser$1.ID);
+                    this.match(MiniParser.ID);
 
                                           localctx.type = 'select_id';
                                       
@@ -13149,17 +13053,17 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 9:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 111;
                     if (!( this.precpred(this._ctx, 18))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 18)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 18)");
                     }
                     this.state = 112;
-                    this.match(MiniParser$1.T__1);
+                    this.match(MiniParser.T__1);
                     this.state = 113;
                     this.expr(0);
                     this.state = 114;
-                    this.match(MiniParser$1.T__2);
+                    this.match(MiniParser.T__2);
 
                                           localctx.type = 'select_expr';
                                       
@@ -13167,13 +13071,13 @@ MiniParser$1.prototype.expr = function(_p) {
 
                 case 10:
                     localctx = new ExprContext(this, _parentctx, _parentState);
-                    this.pushNewRecursionContext(localctx, _startState, MiniParser$1.RULE_expr);
+                    this.pushNewRecursionContext(localctx, _startState, MiniParser.RULE_expr);
                     this.state = 117;
                     if (!( this.precpred(this._ctx, 9))) {
-                        throw new antlr4$1.error.FailedPredicateException(this, "this.precpred(this._ctx, 9)");
+                        throw new _exports.error.FailedPredicateException(this, "this.precpred(this._ctx, 9)");
                     }
                     this.state = 118;
-                    this.match(MiniParser$1.T__18);
+                    this.match(MiniParser.T__18);
                     this.state = 119;
                     this.function_call();
                      localctx.type = 'pipe'; 
@@ -13187,7 +13091,7 @@ MiniParser$1.prototype.expr = function(_p) {
         }
 
     } catch( error) {
-        if(error instanceof antlr4$1.error.RecognitionException) {
+        if(error instanceof _exports.error.RecognitionException) {
 	        localctx.exception = error;
 	        this._errHandler.reportError(this, error);
 	        this._errHandler.recover(this, error);
@@ -13207,19 +13111,19 @@ function Function_callContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_function_call;
-    this.name = null; // Token
-    this.args = null; // Call_argumentsContext
+    this.ruleIndex = MiniParser.RULE_function_call;
+    this.name = null; 
+    this.args = null; 
     return this;
 }
 
-Function_callContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+Function_callContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 Function_callContext.prototype.constructor = Function_callContext;
 
 Function_callContext.prototype.ID = function() {
-    return this.getToken(MiniParser$1.ID, 0);
+    return this.getToken(MiniParser.ID, 0);
 };
 
 Function_callContext.prototype.call_arguments = function() {
@@ -13241,12 +13145,12 @@ Function_callContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.Function_callContext = Function_callContext;
+MiniParser.Function_callContext = Function_callContext;
 
-MiniParser$1.prototype.function_call = function() {
+MiniParser.prototype.function_call = function() {
 
     var localctx = new Function_callContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 6, MiniParser$1.RULE_function_call);
+    this.enterRule(localctx, 6, MiniParser.RULE_function_call);
     try {
         this.state = 136;
         this._errHandler.sync(this);
@@ -13255,28 +13159,28 @@ MiniParser$1.prototype.function_call = function() {
         case 1:
             this.enterOuterAlt(localctx, 1);
             this.state = 127;
-            localctx.name = this.match(MiniParser$1.ID);
+            localctx.name = this.match(MiniParser.ID);
             this.state = 128;
-            this.match(MiniParser$1.T__21);
+            this.match(MiniParser.T__21);
              localctx.type = 'call'; 
             break;
 
         case 2:
             this.enterOuterAlt(localctx, 2);
             this.state = 130;
-            localctx.name = this.match(MiniParser$1.ID);
+            localctx.name = this.match(MiniParser.ID);
             this.state = 131;
-            this.match(MiniParser$1.T__19);
+            this.match(MiniParser.T__19);
             this.state = 132;
             localctx.args = this.call_arguments();
             this.state = 133;
-            this.match(MiniParser$1.T__20);
+            this.match(MiniParser.T__20);
              localctx.type = 'call'; 
             break;
 
         }
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13296,21 +13200,21 @@ function NumberContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_number;
+    this.ruleIndex = MiniParser.RULE_number;
     return this;
 }
 
-NumberContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+NumberContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 NumberContext.prototype.constructor = NumberContext;
 
 NumberContext.prototype.INT = function() {
-    return this.getToken(MiniParser$1.INT, 0);
+    return this.getToken(MiniParser.INT, 0);
 };
 
 NumberContext.prototype.FLOAT = function() {
-    return this.getToken(MiniParser$1.FLOAT, 0);
+    return this.getToken(MiniParser.FLOAT, 0);
 };
 
 NumberContext.prototype.enterRule = function(listener) {
@@ -13328,33 +13232,33 @@ NumberContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.NumberContext = NumberContext;
+MiniParser.NumberContext = NumberContext;
 
-MiniParser$1.prototype.number = function() {
+MiniParser.prototype.number = function() {
 
     var localctx = new NumberContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 8, MiniParser$1.RULE_number);
+    this.enterRule(localctx, 8, MiniParser.RULE_number);
     try {
         this.state = 142;
         this._errHandler.sync(this);
         switch(this._input.LA(1)) {
-        case MiniParser$1.INT:
+        case MiniParser.INT:
             this.enterOuterAlt(localctx, 1);
             this.state = 138;
-            this.match(MiniParser$1.INT);
+            this.match(MiniParser.INT);
              localctx.type = 'int'; 
             break;
-        case MiniParser$1.FLOAT:
+        case MiniParser.FLOAT:
             this.enterOuterAlt(localctx, 2);
             this.state = 140;
-            this.match(MiniParser$1.FLOAT);
+            this.match(MiniParser.FLOAT);
              localctx.type = 'float'; 
             break;
         default:
-            throw new antlr4$1.error.NoViableAltException(this);
+            throw new _exports.error.NoViableAltException(this);
         }
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13374,15 +13278,15 @@ function SeqContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_seq;
-    this._expr = null; // ExprContext
-    this.items = []; // of ExprContexts
+    this.ruleIndex = MiniParser.RULE_seq;
+    this._expr = null; 
+    this.items = []; 
     return this;
 }
 
-SeqContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+SeqContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 SeqContext.prototype.constructor = SeqContext;
 
 SeqContext.prototype.expr = function(i) {
@@ -13411,13 +13315,13 @@ SeqContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.SeqContext = SeqContext;
+MiniParser.SeqContext = SeqContext;
 
-MiniParser$1.prototype.seq = function() {
+MiniParser.prototype.seq = function() {
 
     var localctx = new SeqContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 10, MiniParser$1.RULE_seq);
-    var _la = 0; // Token type
+    this.enterRule(localctx, 10, MiniParser.RULE_seq);
+    var _la = 0; 
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 144;
@@ -13426,9 +13330,9 @@ MiniParser$1.prototype.seq = function() {
         this.state = 149;
         this._errHandler.sync(this);
         _la = this._input.LA(1);
-        while(_la===MiniParser$1.T__22) {
+        while(_la===MiniParser.T__22) {
             this.state = 145;
-            this.match(MiniParser$1.T__22);
+            this.match(MiniParser.T__22);
             this.state = 146;
             localctx._expr = this.expr(0);
             localctx.items.push(localctx._expr);
@@ -13437,7 +13341,7 @@ MiniParser$1.prototype.seq = function() {
             _la = this._input.LA(1);
         }
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13457,15 +13361,15 @@ function Id_seqContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_id_seq;
-    this._ID = null; // Token
-    this.items = []; // of Tokens
+    this.ruleIndex = MiniParser.RULE_id_seq;
+    this._ID = null; 
+    this.items = []; 
     return this;
 }
 
-Id_seqContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+Id_seqContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 Id_seqContext.prototype.constructor = Id_seqContext;
 
 Id_seqContext.prototype.ID = function(i) {
@@ -13473,9 +13377,9 @@ Id_seqContext.prototype.ID = function(i) {
 		i = null;
 	}
     if(i===null) {
-        return this.getTokens(MiniParser$1.ID);
+        return this.getTokens(MiniParser.ID);
     } else {
-        return this.getToken(MiniParser$1.ID, i);
+        return this.getToken(MiniParser.ID, i);
     }
 };
 
@@ -13495,33 +13399,33 @@ Id_seqContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.Id_seqContext = Id_seqContext;
+MiniParser.Id_seqContext = Id_seqContext;
 
-MiniParser$1.prototype.id_seq = function() {
+MiniParser.prototype.id_seq = function() {
 
     var localctx = new Id_seqContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 12, MiniParser$1.RULE_id_seq);
-    var _la = 0; // Token type
+    this.enterRule(localctx, 12, MiniParser.RULE_id_seq);
+    var _la = 0; 
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 152;
-        localctx._ID = this.match(MiniParser$1.ID);
+        localctx._ID = this.match(MiniParser.ID);
         localctx.items.push(localctx._ID);
         this.state = 157;
         this._errHandler.sync(this);
         _la = this._input.LA(1);
-        while(_la===MiniParser$1.T__22) {
+        while(_la===MiniParser.T__22) {
             this.state = 153;
-            this.match(MiniParser$1.T__22);
+            this.match(MiniParser.T__22);
             this.state = 154;
-            localctx._ID = this.match(MiniParser$1.ID);
+            localctx._ID = this.match(MiniParser.ID);
             localctx.items.push(localctx._ID);
             this.state = 159;
             this._errHandler.sync(this);
             _la = this._input.LA(1);
         }
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13541,15 +13445,15 @@ function Call_argumentsContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_call_arguments;
-    this.args = null; // Positional_argumentsContext
-    this.namedArgs = null; // Named_argumentsContext
+    this.ruleIndex = MiniParser.RULE_call_arguments;
+    this.args = null; 
+    this.namedArgs = null; 
     return this;
 }
 
-Call_argumentsContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+Call_argumentsContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 Call_argumentsContext.prototype.constructor = Call_argumentsContext;
 
 Call_argumentsContext.prototype.positional_arguments = function() {
@@ -13575,12 +13479,12 @@ Call_argumentsContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.Call_argumentsContext = Call_argumentsContext;
+MiniParser.Call_argumentsContext = Call_argumentsContext;
 
-MiniParser$1.prototype.call_arguments = function() {
+MiniParser.prototype.call_arguments = function() {
 
     var localctx = new Call_argumentsContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 14, MiniParser$1.RULE_call_arguments);
+    this.enterRule(localctx, 14, MiniParser.RULE_call_arguments);
     try {
         this.state = 167;
         this._errHandler.sync(this);
@@ -13608,14 +13512,14 @@ MiniParser$1.prototype.call_arguments = function() {
             this.state = 163;
             localctx.args = this.positional_arguments();
             this.state = 164;
-            this.match(MiniParser$1.T__22);
+            this.match(MiniParser.T__22);
             this.state = 165;
             localctx.namedArgs = this.named_arguments();
             break;
 
         }
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13635,13 +13539,13 @@ function Positional_argumentsContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_positional_arguments;
+    this.ruleIndex = MiniParser.RULE_positional_arguments;
     return this;
 }
 
-Positional_argumentsContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+Positional_argumentsContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 Positional_argumentsContext.prototype.constructor = Positional_argumentsContext;
 
 Positional_argumentsContext.prototype.expr = function(i) {
@@ -13670,13 +13574,13 @@ Positional_argumentsContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.Positional_argumentsContext = Positional_argumentsContext;
+MiniParser.Positional_argumentsContext = Positional_argumentsContext;
 
-MiniParser$1.prototype.positional_arguments = function() {
+MiniParser.prototype.positional_arguments = function() {
 
     var localctx = new Positional_argumentsContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 16, MiniParser$1.RULE_positional_arguments);
-    var _la = 0; // Token type
+    this.enterRule(localctx, 16, MiniParser.RULE_positional_arguments);
+    var _la = 0; 
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 169;
@@ -13684,14 +13588,14 @@ MiniParser$1.prototype.positional_arguments = function() {
         this.state = 176;
         this._errHandler.sync(this);
         var _alt = this._interp.adaptivePredict(this._input,10,this._ctx);
-        while(_alt!=2 && _alt!=antlr4$1.atn.ATN.INVALID_ALT_NUMBER) {
+        while(_alt!=2 && _alt!=_exports.atn.ATN.INVALID_ALT_NUMBER) {
             if(_alt===1) {
                 this.state = 170;
-                this.match(MiniParser$1.T__22);
+                this.match(MiniParser.T__22);
                 this.state = 172;
                 this._errHandler.sync(this);
                 _la = this._input.LA(1);
-                if((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser$1.T__1) | (1 << MiniParser$1.T__3) | (1 << MiniParser$1.T__4) | (1 << MiniParser$1.T__5) | (1 << MiniParser$1.T__19) | (1 << MiniParser$1.T__23) | (1 << MiniParser$1.BOOLEAN) | (1 << MiniParser$1.ID) | (1 << MiniParser$1.INT) | (1 << MiniParser$1.FLOAT) | (1 << MiniParser$1.STRING))) !== 0)) {
+                if((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser.T__1) | (1 << MiniParser.T__3) | (1 << MiniParser.T__4) | (1 << MiniParser.T__5) | (1 << MiniParser.T__19) | (1 << MiniParser.T__23) | (1 << MiniParser.BOOLEAN) | (1 << MiniParser.ID) | (1 << MiniParser.INT) | (1 << MiniParser.FLOAT) | (1 << MiniParser.STRING))) !== 0)) {
                     this.state = 171;
                     this.expr(0);
                 }
@@ -13703,7 +13607,7 @@ MiniParser$1.prototype.positional_arguments = function() {
         }
 
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13723,13 +13627,13 @@ function Named_argumentsContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_named_arguments;
+    this.ruleIndex = MiniParser.RULE_named_arguments;
     return this;
 }
 
-Named_argumentsContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+Named_argumentsContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 Named_argumentsContext.prototype.constructor = Named_argumentsContext;
 
 Named_argumentsContext.prototype.named_argument = function(i) {
@@ -13758,13 +13662,13 @@ Named_argumentsContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.Named_argumentsContext = Named_argumentsContext;
+MiniParser.Named_argumentsContext = Named_argumentsContext;
 
-MiniParser$1.prototype.named_arguments = function() {
+MiniParser.prototype.named_arguments = function() {
 
     var localctx = new Named_argumentsContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 18, MiniParser$1.RULE_named_arguments);
-    var _la = 0; // Token type
+    this.enterRule(localctx, 18, MiniParser.RULE_named_arguments);
+    var _la = 0; 
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 179;
@@ -13772,13 +13676,13 @@ MiniParser$1.prototype.named_arguments = function() {
         this.state = 186;
         this._errHandler.sync(this);
         _la = this._input.LA(1);
-        while(_la===MiniParser$1.T__22) {
+        while(_la===MiniParser.T__22) {
             this.state = 180;
-            this.match(MiniParser$1.T__22);
+            this.match(MiniParser.T__22);
             this.state = 182;
             this._errHandler.sync(this);
             _la = this._input.LA(1);
-            if(_la===MiniParser$1.ID) {
+            if(_la===MiniParser.ID) {
                 this.state = 181;
                 this.named_argument();
             }
@@ -13788,7 +13692,7 @@ MiniParser$1.prototype.named_arguments = function() {
             _la = this._input.LA(1);
         }
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13808,21 +13712,21 @@ function Named_argumentContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_named_argument;
+    this.ruleIndex = MiniParser.RULE_named_argument;
     return this;
 }
 
-Named_argumentContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+Named_argumentContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 Named_argumentContext.prototype.constructor = Named_argumentContext;
 
 Named_argumentContext.prototype.ID = function() {
-    return this.getToken(MiniParser$1.ID, 0);
+    return this.getToken(MiniParser.ID, 0);
 };
 
 Named_argumentContext.prototype.EQ = function() {
-    return this.getToken(MiniParser$1.EQ, 0);
+    return this.getToken(MiniParser.EQ, 0);
 };
 
 Named_argumentContext.prototype.expr = function() {
@@ -13844,23 +13748,23 @@ Named_argumentContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.Named_argumentContext = Named_argumentContext;
+MiniParser.Named_argumentContext = Named_argumentContext;
 
-MiniParser$1.prototype.named_argument = function() {
+MiniParser.prototype.named_argument = function() {
 
     var localctx = new Named_argumentContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 20, MiniParser$1.RULE_named_argument);
+    this.enterRule(localctx, 20, MiniParser.RULE_named_argument);
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 189;
-        this.match(MiniParser$1.ID);
+        this.match(MiniParser.ID);
         this.state = 190;
-        this.match(MiniParser$1.EQ);
+        this.match(MiniParser.EQ);
         this.state = 191;
         this.expr(0);
          localctx.type = 'named-argument'; 
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13880,13 +13784,13 @@ function ArrayContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_array;
+    this.ruleIndex = MiniParser.RULE_array;
     return this;
 }
 
-ArrayContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+ArrayContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 ArrayContext.prototype.constructor = ArrayContext;
 
 ArrayContext.prototype.seq = function() {
@@ -13908,30 +13812,30 @@ ArrayContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.ArrayContext = ArrayContext;
+MiniParser.ArrayContext = ArrayContext;
 
-MiniParser$1.prototype.array = function() {
+MiniParser.prototype.array = function() {
 
     var localctx = new ArrayContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 22, MiniParser$1.RULE_array);
-    var _la = 0; // Token type
+    this.enterRule(localctx, 22, MiniParser.RULE_array);
+    var _la = 0; 
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 194;
-        this.match(MiniParser$1.T__1);
+        this.match(MiniParser.T__1);
         this.state = 196;
         this._errHandler.sync(this);
         _la = this._input.LA(1);
-        if((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser$1.T__1) | (1 << MiniParser$1.T__3) | (1 << MiniParser$1.T__4) | (1 << MiniParser$1.T__5) | (1 << MiniParser$1.T__19) | (1 << MiniParser$1.T__23) | (1 << MiniParser$1.BOOLEAN) | (1 << MiniParser$1.ID) | (1 << MiniParser$1.INT) | (1 << MiniParser$1.FLOAT) | (1 << MiniParser$1.STRING))) !== 0)) {
+        if((((_la) & ~0x1f) == 0 && ((1 << _la) & ((1 << MiniParser.T__1) | (1 << MiniParser.T__3) | (1 << MiniParser.T__4) | (1 << MiniParser.T__5) | (1 << MiniParser.T__19) | (1 << MiniParser.T__23) | (1 << MiniParser.BOOLEAN) | (1 << MiniParser.ID) | (1 << MiniParser.INT) | (1 << MiniParser.FLOAT) | (1 << MiniParser.STRING))) !== 0)) {
             this.state = 195;
             this.seq();
         }
 
         this.state = 198;
-        this.match(MiniParser$1.T__2);
+        this.match(MiniParser.T__2);
          localctx.type = 'array'; 
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -13951,17 +13855,17 @@ function ObjectContext(parser, parent, invokingState) {
 	if(invokingState===undefined || invokingState===null) {
 		invokingState = -1;
 	}
-	antlr4$1.ParserRuleContext.call(this, parent, invokingState);
+	_exports.ParserRuleContext.call(this, parent, invokingState);
     this.parser = parser;
-    this.ruleIndex = MiniParser$1.RULE_object;
-    this._ID = null; // Token
-    this.keys = []; // of Tokens
-    this._expr = null; // ExprContext
-    this.vals = []; // of ExprContexts
+    this.ruleIndex = MiniParser.RULE_object;
+    this._ID = null; 
+    this.keys = []; 
+    this._expr = null; 
+    this.vals = []; 
     return this;
 }
 
-ObjectContext.prototype = Object.create(antlr4$1.ParserRuleContext.prototype);
+ObjectContext.prototype = Object.create(_exports.ParserRuleContext.prototype);
 ObjectContext.prototype.constructor = ObjectContext;
 
 ObjectContext.prototype.ID = function(i) {
@@ -13969,9 +13873,9 @@ ObjectContext.prototype.ID = function(i) {
 		i = null;
 	}
     if(i===null) {
-        return this.getTokens(MiniParser$1.ID);
+        return this.getTokens(MiniParser.ID);
     } else {
-        return this.getToken(MiniParser$1.ID, i);
+        return this.getToken(MiniParser.ID, i);
     }
 };
 
@@ -14002,40 +13906,40 @@ ObjectContext.prototype.exitRule = function(listener) {
 
 
 
-MiniParser$1.ObjectContext = ObjectContext;
+MiniParser.ObjectContext = ObjectContext;
 
-MiniParser$1.prototype.object = function() {
+MiniParser.prototype.object = function() {
 
     var localctx = new ObjectContext(this, this._ctx, this.state);
-    this.enterRule(localctx, 24, MiniParser$1.RULE_object);
-    var _la = 0; // Token type
+    this.enterRule(localctx, 24, MiniParser.RULE_object);
+    var _la = 0; 
     try {
         this.enterOuterAlt(localctx, 1);
         this.state = 201;
-        this.match(MiniParser$1.T__23);
+        this.match(MiniParser.T__23);
         this.state = 214;
         this._errHandler.sync(this);
         _la = this._input.LA(1);
-        if(_la===MiniParser$1.ID) {
+        if(_la===MiniParser.ID) {
             this.state = 202;
-            localctx._ID = this.match(MiniParser$1.ID);
+            localctx._ID = this.match(MiniParser.ID);
             localctx.keys.push(localctx._ID);
             this.state = 203;
-            this.match(MiniParser$1.T__24);
+            this.match(MiniParser.T__24);
             this.state = 204;
             localctx._expr = this.expr(0);
             localctx.vals.push(localctx._expr);
             this.state = 211;
             this._errHandler.sync(this);
             _la = this._input.LA(1);
-            while(_la===MiniParser$1.T__22) {
+            while(_la===MiniParser.T__22) {
                 this.state = 205;
-                this.match(MiniParser$1.T__22);
+                this.match(MiniParser.T__22);
                 this.state = 206;
-                localctx._ID = this.match(MiniParser$1.ID);
+                localctx._ID = this.match(MiniParser.ID);
                 localctx.keys.push(localctx._ID);
                 this.state = 207;
-                this.match(MiniParser$1.T__24);
+                this.match(MiniParser.T__24);
                 this.state = 208;
                 localctx._expr = this.expr(0);
                 localctx.vals.push(localctx._expr);
@@ -14046,10 +13950,10 @@ MiniParser$1.prototype.object = function() {
         }
 
         this.state = 216;
-        this.match(MiniParser$1.T__25);
+        this.match(MiniParser.T__25);
          localctx.type = 'object'; 
     } catch (re) {
-    	if(re instanceof antlr4$1.error.RecognitionException) {
+    	if(re instanceof _exports.error.RecognitionException) {
 	        localctx.exception = re;
 	        this._errHandler.reportError(this, re);
 	        this._errHandler.recover(this, re);
@@ -14063,7 +13967,7 @@ MiniParser$1.prototype.object = function() {
 };
 
 
-MiniParser$1.prototype.sempred = function(localctx, ruleIndex, predIndex) {
+MiniParser.prototype.sempred = function(localctx, ruleIndex, predIndex) {
 	switch(ruleIndex) {
 	case 2:
 			return this.expr_sempred(localctx, predIndex);
@@ -14072,10 +13976,10 @@ MiniParser$1.prototype.sempred = function(localctx, ruleIndex, predIndex) {
    }
 };
 
-MiniParser$1.prototype.expr_sempred = function(localctx, predIndex) {
+MiniParser.prototype.expr_sempred = function(localctx, predIndex) {
 	switch(predIndex) {
 		case 0:
-			return this._input.LA(1) !== MiniParser$1.EQ;
+			return this._input.LA(1) !== MiniParser.EQ;
 		case 1:
 			return this.precpred(this._ctx, 16);
 		case 2:
@@ -14102,28 +14006,16 @@ MiniParser$1.prototype.expr_sempred = function(localctx, predIndex) {
 };
 
 
-var MiniParser_2 = MiniParser$1;
+var MiniParser_2 = MiniParser;
 
 var MiniParser_1 = {
 	MiniParser: MiniParser_2
 };
 
-var InputStream = antlr4$1.InputStream;
-var CommonTokenStream = antlr4$1.CommonTokenStream;
-var MiniLexer = MiniLexer_1.MiniLexer;
-var MiniParser = MiniParser_1.MiniParser;
-var MiniListener = MiniListener_1.MiniListener;
-var treeWalker = antlr4$1.tree.ParseTreeWalker.DEFAULT;
+const { InputStream, CommonTokenStream } = _exports;
+const treeWalker = _exports.tree.ParseTreeWalker.DEFAULT;
+const MiniLexer$1 = MiniLexer_1.MiniLexer;
+const MiniParser$1 = MiniParser_1.MiniParser;
+const MiniListener$2 = MiniListener_1.MiniListener;
 
-var _parser = {
-	InputStream: InputStream,
-	CommonTokenStream: CommonTokenStream,
-	MiniLexer: MiniLexer,
-	MiniParser: MiniParser,
-	MiniListener: MiniListener,
-	treeWalker: treeWalker
-};
-
-export { InputStream, CommonTokenStream, MiniLexer, MiniParser, MiniListener, treeWalker };export default _parser;
-
-//# sourceMappingURL=./parser.js.map
+export { InputStream, CommonTokenStream, treeWalker, MiniLexer$1 as MiniLexer, MiniParser$1 as MiniParser, MiniListener$2 as MiniListener };
